updating .env.example file
