+++ .env.example
@@ -27,7 +27,6 @@ CLIENT_ID=google_client_id
 CLIENT_SECRET=google_client_secret
 REDIRECT_URL=google_REDIRECT_URL
 
-
 # cloudinary
 
 CLOUD_NAME=
