+++ .sequelizerc
@@ -2,7 +2,7 @@ const path= require('path');
 
 module.exports = {
   config: path.resolve('./src/config', 'config.js'),
-  'models-path': path.resolve('./src/models'),
+  'models-path': path.resolve('./src/database/models'),
   'seeders-path': path.resolve('./src/database/seeders'),
   'migrations-path': path.resolve('./src/database/migrations'),
 }
