+++ src/controllers/products/wishlistController.js
@@ -1,6 +1,7 @@
-import db from '../../models';
+import db from '../../database/models';
 import { Op } from 'sequelize';
 import { asyncWrapper } from '../../utils/handlingTryCatchBlocks';
+import { checkEmptyFields } from '../../utils/validations/handlingEmptyFields';
 export const getWishlist = asyncWrapper(async (req, res) => {
   const wishlist = await db.wishlist.findOne({
     where: { userId: req.user.id },
@@ -17,8 +18,10 @@ export const getWishlist = asyncWrapper(async (req, res) => {
   });
   res.status(200).json({ status: req.t('success'), products });
 });
-export const  addToWishlist = asyncWrapper( async(req, res) => {
-    const product = req.body.productId;
+export const addToWishlist = asyncWrapper(async (req, res) => {
+  const product = req.body.productId;
+  const emptyFields = checkEmptyFields(req.body);
+  if (!emptyFields) {
     const activewishlist = await db.wishlist.findOne({
       where: {
         userId: req.user.id,
@@ -33,7 +36,7 @@ export const  addToWishlist = asyncWrapper( async(req, res) => {
       });
       res
         .status(201)
-        .json({status: req.t('success'), message: req.t('wishlist_created') });
+        .json({ status: req.t('success'), message: req.t('wishlist_created') });
     } else {
       //updating the wishlist when its there
       if (activewishlist.productId.includes(product)) {
@@ -43,7 +46,6 @@ export const  addToWishlist = asyncWrapper( async(req, res) => {
         });
       } else {
         const updatedWishlist = [...activewishlist.productId, product];
-        console.log(activewishlist.productId);
         await activewishlist.update({
           productId: updatedWishlist,
         });
@@ -53,30 +55,36 @@ export const  addToWishlist = asyncWrapper( async(req, res) => {
         });
       }
     }
-})
-export const  deleteFromWishlist = asyncWrapper(async(req, res)=>{
-    const product = req.params.product;
-    const activewishlist = await db.wishlist.findOne({
-      where: {
-        userId: req.user.id,
-      },
+  }
+});
+export const deleteFromWishlist = asyncWrapper(async (req, res) => {
+  const product = req.params.product;
+  const activewishlist = await db.wishlist.findOne({
+    where: {
+      userId: req.user.id,
+    },
+  });
+  const newwishlist = [...activewishlist.productId];
+  if (newwishlist.includes(product)) {
+    newwishlist.filter((element, index) => {
+      if (element === product) {
+        newwishlist.splice(index, 1);
+      }
     });
-    const newwishlist = [...activewishlist.productId];
-    if (newwishlist.includes(product)) {
-      newwishlist.filter((element, index) => {
-        if (element === product) {
-          newwishlist.splice(index, 1);
-        }
-      });
 
-      await activewishlist.update({
-        productId: newwishlist,
+    await activewishlist.update({
+      productId: newwishlist,
+    });
+    res
+      .status(200)
+      .json({
+        status: req.t('success'),
+        message: req.t('removed_from_wishlist'),
       });
-      res.status(200).json({status:req.t('success'), message: req.t('removed_from_wishlist') });
-    } else {
-      return res.status(400).json({
-        status: req.t('fail'),
-        message: req.t('not_exists_in_wishlist'),
-      });
-    }
-})
+  } else {
+    return res.status(400).json({
+      status: req.t('fail'),
+      message: req.t('not_exists_in_wishlist'),
+    });
+  }
+});
