+++ src/routes/products/buyer.routes.js
@@ -6,7 +6,7 @@ import {
 import { auth, isUserEnabled } from '../../middleware/auth';
 const routes = express.Router();
 
-routes.get('/:id', auth('buyer'), isUserEnabled, viewSingleProduct)
+routes.get('/:id', viewSingleProduct)
 
 routes.get('/view/recommendations',auth('buyer'),recommendedProducts);
