+++ src/routes/products/category.js
