+++ src/utils/verifyPassword.js
@@ -10,7 +10,6 @@ export const signToken = async (user) => {
   let expires = process.env.JWT_EXPIRES_IN;
   if(user.reason === 'reset'){
     expires = process.env.PASS_EXPIRES_IN;
-    // console.log(user);
   }
   const result = await jwt.sign({ user }, process.env.JWT_SECRET, {
     expiresIn: expires,
